Updated db and endpoints / apis
learning a lot about how express routing works.

Updated users to include their primary language to be used later when we get to translating.

Made another collection for messages that stores the message, the origin user, the destination user, and the message time/date - with this info, the previous chats can be populated in the front end.